+++ src/base/i-base-mixin-conf.ts
@@ -1,6 +1,7 @@
-import {MinimalCFDimension, MinimalCFGroup} from '../core/types';
+import {BaseAccessor, MinimalCFDimension} from '../core/types';
 
 export interface IBaseMixinConf {
+    ordering?: BaseAccessor<any>;
     dimension?: MinimalCFDimension;
     useViewBoxResizing?: boolean;
     minWidth?: number;
